move useDelete into own file
