+++ client/packages/system/src/Asset/api/hooks/index.ts
@@ -3,6 +3,7 @@ export * from './useAssetList';
 export * from './useAssetInsert';
 export * from './useAssetLogReasonList';
 export * from './useAssetLogReason';
+export * from './useAssetDelete';
 
 export const useAssetData = {
   utils: {
